ProcessValidator: (created a seperate process validator class)
